some change
Rewrite JS code to make the <p> elements is only created once.
Use watchPosition(), so it will repeatedly updating the location value. (automatically) then, it will do what ever the code is written in its call back function (give that <p> element with id="location" the json string with latitude and longitude value)
We only need to execute the function once, watchPosition() will execute repeatedly and updating the new value for latitude and longitude if necessary.
That's why, in the loop, we only need to read the string from the <p> element, and don't need to call the function again.